linting and python version fixes
